Small fixes/workarounds for Apple Xcode
The flex shipped with Apple Xcode has an issue with signed/unsigned
comparison mismatch. Don't make this an error for the generated
pgnscanner.cc.

Also, the signature of LexerInput() is slightly different, so change
the signature of the currently implemented function and provide a
wrapper for compatibility with upstream flex.